adding files to folder
