--- index.php
@@ -1,6 +1,0 @@
-<?php
-  // For now, index.php just redirects to browse.php, but you can change this
-  // if you like.
-  
-  header("Location: browse.php");
-?>
