--- logout.php
@@ -1,14 +1,0 @@
-<?php
-
-session_start();
-
-unset($_SESSION['logged_in']);
-unset($_SESSION['account_type']);
-setcookie(session_name(), "", time() - 360);
-session_destroy();
-
-
-// Redirect to index
-header("Location: index.php");
-
-?>
