--- place_bid.php
@@ -1,6 +1,0 @@
-<?php
-
-// TODO: Extract $_POST variables, check they're OK, and attempt to make a bid.
-// Notify user of success/failure and redirect/give navigation options.
-
-?>
