--- process_registration.php
@@ -1,7 +1,0 @@
-<?php
-
-// TODO: Extract $_POST variables, check they're OK, and attempt to create
-// an account. Notify user of success/failure and redirect/give navigation 
-// options.
-
-?>
