+++ starter-code/login_result.php
@@ -1,0 +1,18 @@
+<?php
+
+// TODO: Extract $_POST variables, check they're OK, and attempt to login.
+// Notify user of success/failure and redirect/give navigation options.
+
+// For now, I will just set session variables and redirect.
+
+session_start();
+$_SESSION['logged_in'] = true;
+$_SESSION['username'] = "test";
+$_SESSION['account_type'] = "buyer";
+
+echo('<div class="text-center">You are now logged in! You will be redirected shortly.</div>');
+
+// Redirect to index after 5 seconds
+header("refresh:5;url=index.php");
+
+?>
