+++ starter-code/mybids.php
@@ -1,0 +1,24 @@
+<?php include_once("header.php")?>
+<?php require("utilities.php")?>
+
+<div class="container">
+
+<h2 class="my-3">My bids</h2>
+
+<?php
+  // This page is for showing a user the auctions they've bid on.
+  // It will be pretty similar to browse.php, except there is no search bar.
+  // This can be started after browse.php is working with a database.
+  // Feel free to extract out useful functions from browse.php and put them in
+  // the shared "utilities.php" where they can be shared by multiple files.
+  
+  
+  // TODO: Check user's credentials (cookie/session).
+  
+  // TODO: Perform a query to pull up the auctions they've bidded on.
+  
+  // TODO: Loop through results and print them out as list items.
+  
+?>
+
+<?php include_once("footer.php")?>
