+++ starter-code/recommendations.php
@@ -1,0 +1,22 @@
+<?php include_once("header.php")?>
+<?php require("utilities.php")?>
+
+<div class="container">
+
+<h2 class="my-3">Recommendations for you</h2>
+
+<?php
+  // This page is for showing a buyer recommended items based on their bid 
+  // history. It will be pretty similar to browse.php, except there is no 
+  // search bar. This can be started after browse.php is working with a database.
+  // Feel free to extract out useful functions from browse.php and put them in
+  // the shared "utilities.php" where they can be shared by multiple files.
+  
+  
+  // TODO: Check user's credentials (cookie/session).
+  
+  // TODO: Perform a query to pull up auctions they might be interested in.
+  
+  // TODO: Loop through results and print them out as list items.
+  
+?>
